Test interpreter
